requirements.txt
streamlit
Wikipedia
speechrecognition
googlesearch-python
requests
beautifulsoup4
